+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+
+### 🚀 Features
+
+- *(deploy)* Add service unit names to template context
+- *(app)* Add --force flag to restart command
+
 ## [0.20.7] - 2026-02-03
 
 ### 🚀 Features